+++ wizwalker/memory/memory_objects/combat_action.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from wizwalker.memory.memanagers import MemoryView
-from wizwalker.memory.memtypes import *
+from wizwalker.memory.memonster.memanagers import MemoryView
+from wizwalker.memory.memonster.memtypes import *
 from .spell import Spell
