+++ wizwalker/memory/memory_objects/combat_participant.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from wizwalker.memory.memory_object import PropertyClass
-from wizwalker.memory.memtypes import *
+from wizwalker.memory.memonster.memtypes import *
 from .enums import PipAquiredByEnum
 from .game_stats import GameStats
 from .spell import Hand
