+++ wizwalker/memory/memory_objects/combat_resolver.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from wizwalker.memory.memory_object import PropertyClass
-from wizwalker.memory.memtypes import *
+from wizwalker.memory.memonster.memtypes import *
 from .spell_effect import SpellEffect
